feat: agora é possivel adicionar produtos ao carrinho e gerar novos p…

+++ src/main/java/br/com/tech/springtech/domain/model/ProdutoCarrinho.java
@@ -3,7 +3,6 @@ package br.com.tech.springtech.domain.model;
 import java.math.BigDecimal;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
-
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -22,7 +21,7 @@ public class ProdutoCarrinho {
     @EqualsAndHashCode.Include
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer produtoCarrinhoId;
+    private Long produtoCarrinhoId;
 
     @Column(nullable = false)
     private Long quantidade;
